Permission Ranks Table
I have noticed people have been frequently confused by the permissions system the way it is laid out as the rank description is under the actual rank. So, I formatted them to be side by side in a table so it may be more easily understood. Included a quick refinement of the notes bullet including the default dj rank, and a quick grammar check at the end.